feat: implement user balance handling with BigDecimal, add auction subscription events, and introduce UserFactory for improved object rehydration. Introduce BroadcastBroker

+++ client/src/main/java/com/auction/client/page/admin/AdminPageController.java
@@ -192,7 +192,7 @@ public class AdminPageController implements LifecycleAwareController {
                                                             "$"
                                                                     + MONEY_FMT.format(
                                                                             u.getBalance() != null
-                                                                                    ? u.getBalance()
+                                                                                    ? u.getBalance().doubleValue()
                                                                                     : 0.0),
                                                             "Active"));
                                         }
